raise if stream change fails
