+++ polyn_elixir_client/lib/polyn/stream_migrator/stream_migrator_exception.ex
@@ -1,0 +1,6 @@
+defmodule Polyn.StreamMigrator.Exception do
+  @moduledoc """
+  Exception when changing the configuration of streams and consumers
+  """
+  defexception [:message]
+end
